# More efficient implementation of `Optional.OptionalImpl`