Below are results before and after with Scala 2.13 and JDK 21 on Intel® Core™ i7-11800H.

Before:
```
[info] Benchmark                             (size)   Mode  Cnt          Score          Error  Units
[info] OptionalGetOptionBenchmark.direct        N/A  thrpt    5  420172354.268 ±  6533346.984  ops/s
[info] OptionalGetOptionBenchmark.monocle       N/A  thrpt    5   25445616.822 ±   281682.766  ops/s
[info] OptionalGetOptionBenchmark.zioBlocks     N/A  thrpt    5   12522523.660 ±   157785.797  ops/s
[info] OptionalReplaceBenchmark.direct          N/A  thrpt    5  146784279.377 ±  1873821.225  ops/s
[info] OptionalReplaceBenchmark.monocle         N/A  thrpt    5   15952226.241 ±   477667.121  ops/s
[info] OptionalReplaceBenchmark.quicklens       N/A  thrpt    5   19881390.163 ±   522502.032  ops/s
[info] OptionalReplaceBenchmark.zioBlocks       N/A  thrpt    5    7846006.179 ±    62467.463  ops/s
```

After:
```
[info] Benchmark                             (size)   Mode  Cnt          Score          Error  Units
[info] OptionalGetOptionBenchmark.direct        N/A  thrpt    5  419628685.384 ±  1476396.557  ops/s
[info] OptionalGetOptionBenchmark.monocle       N/A  thrpt    5   25735288.785 ±   304570.517  ops/s
[info] OptionalGetOptionBenchmark.zioBlocks     N/A  thrpt    5   23760660.883 ±   261741.780  ops/s
[info] OptionalReplaceBenchmark.direct          N/A  thrpt    5  146989409.295 ±  2503734.860  ops/s
[info] OptionalReplaceBenchmark.monocle         N/A  thrpt    5   16075406.621 ±   111480.730  ops/s
[info] OptionalReplaceBenchmark.quicklens       N/A  thrpt    5   20399125.020 ±   202409.806  ops/s
[info] OptionalReplaceBenchmark.zioBlocks       N/A  thrpt    5   13606711.542 ±   120201.723  ops/s
```